Renaming query.model.builder to querqy.model.convert #123
