--- querqy-core/src/main/java/querqy/model/builder/QueryBuilderException.java
@@ -1,12 +1,0 @@
-package querqy.model.builder;
-
-public class QueryBuilderException extends RuntimeException {
-
-    public QueryBuilderException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-
-    public QueryBuilderException(final String message) {
-        super(message);
-    }
-}
