--- querqy-core/src/main/java/querqy/model/builder/converter/MapValueConverter.java
@@ -1,8 +1,0 @@
-package querqy.model.builder.converter;
-
-@FunctionalInterface
-public interface MapValueConverter {
-
-    Object toMapValue(final Object builderValue, final MapConverter mapValueConverter);
-
-}
