--- querqy-core/src/main/java/querqy/model/builder/model/DisjunctionMaxClauseBuilder.java
@@ -1,12 +1,0 @@
-package querqy.model.builder.model;
-
-import querqy.model.DisjunctionMaxClause;
-import querqy.model.DisjunctionMaxQuery;
-
-public interface DisjunctionMaxClauseBuilder<B extends QueryNodeBuilder, O extends DisjunctionMaxClause>
-        extends QueryNodeBuilder<B, O, DisjunctionMaxQuery> {
-
-    default DisjunctionMaxClause buildDisjunctionMaxClause(final DisjunctionMaxQuery parent) {
-        return build(parent);
-    }
-}
