--- querqy-core/src/main/java/querqy/model/builder/model/QuerqyQueryBuilder.java
@@ -1,12 +1,0 @@
-package querqy.model.builder.model;
-
-import querqy.model.QuerqyQuery;
-
-public interface QuerqyQueryBuilder<B extends QueryNodeBuilder, O extends QuerqyQuery, P> extends
-        QueryNodeBuilder<B, O, P> {
-
-    default QuerqyQuery<?> buildQuerqyQuery() {
-        return build();
-    }
-
-}
