+++ querqy-core/src/main/java/querqy/model/convert/QueryBuilderException.java
@@ -1,0 +1,12 @@
+package querqy.model.convert;
+
+public class QueryBuilderException extends RuntimeException {
+
+    public QueryBuilderException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+
+    public QueryBuilderException(final String message) {
+        super(message);
+    }
+}
