+++ querqy-core/src/main/java/querqy/model/convert/converter/MapValueConverter.java
@@ -1,0 +1,8 @@
+package querqy.model.convert.converter;
+
+@FunctionalInterface
+public interface MapValueConverter {
+
+    Object toMapValue(final Object builderValue, final MapConverter mapValueConverter);
+
+}
