+++ querqy-core/src/main/java/querqy/model/convert/model/DisjunctionMaxClauseBuilder.java
@@ -1,0 +1,12 @@
+package querqy.model.convert.model;
+
+import querqy.model.DisjunctionMaxClause;
+import querqy.model.DisjunctionMaxQuery;
+
+public interface DisjunctionMaxClauseBuilder<B extends QueryNodeBuilder, O extends DisjunctionMaxClause>
+        extends QueryNodeBuilder<B, O, DisjunctionMaxQuery> {
+
+    default DisjunctionMaxClause buildDisjunctionMaxClause(final DisjunctionMaxQuery parent) {
+        return build(parent);
+    }
+}
