+++ querqy-core/src/main/java/querqy/model/convert/model/QuerqyQueryBuilder.java
@@ -1,0 +1,12 @@
+package querqy.model.convert.model;
+
+import querqy.model.QuerqyQuery;
+
+public interface QuerqyQueryBuilder<B extends QueryNodeBuilder, O extends QuerqyQuery, P> extends
+        QueryNodeBuilder<B, O, P> {
+
+    default QuerqyQuery<?> buildQuerqyQuery() {
+        return build();
+    }
+
+}
