+++ querqy-core/src/test/java/querqy/rewrite/commonrules/model/DecorateInstructionTest.java
@@ -4,7 +4,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.collection.IsIterableContainingInAnyOrder.containsInAnyOrder;
-import static querqy.model.builder.impl.BooleanQueryBuilder.bq;
+import static querqy.model.convert.builder.BooleanQueryBuilder.bq;
 import static querqy.rewrite.commonrules.select.SelectionStrategyFactory.DEFAULT_SELECTION_STRATEGY;
 
 import java.util.AbstractMap;
